backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
langchain==0.0.330
langchain-groq==0.0.1
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
pymupdf==1.23.5
python-docx==0.8.11
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.26.0
pypdf==3.17.0
pytest==7.4.3
httpx==0.25.1
